requirements.txt
streamlit>=1.10
tensorflow>=2.10
numpy
pillow
